dnyarri/mmrnmhrm: Update NixOS configurations.
This is a somewhat messy change, as I experimented around with the new audio
setup, had a filesystem corruption in-between, checked if the memory is okay,
repaired the filesystem, added NIX_OTHER_STORES, found out that it doesn't work
so well, as we might have impurities, proceeded with the audio stuff.

So, that's how both configurations are now and I still need to clean up this
configuration mess by having one single configuration "library" which is shared
by both machines.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>