Add expressions from ~/.nixpkgs/config.nix.
Luckily, this file is already common on all my machines, so there's no need to
consolidate anything, except that it may be useful to set nixpkgs.config to
those expressions as well.

The expressions in this new file are partly unfinished and not yet cleaned up
packages I'm either working on or just needed "right now", so these will
eventually become part of upstream nixpkgs.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>