mmrnmhrm: Add synergy call and configuration.
So, now we no langer need to issue synergys by the window manager or in the
shell, just log in and everything is set up :-)

Well, of course you should only do something like this in a trusted environment,
because this means, that mouse movements and keystrokes are sent unencrypted! Be
sure to set up a SSH tunnel or something similar if you're in a hostile
environment.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>