dnyarri: Add "aszlig" to the audio group.
Pulse doesn't seem to get the right permissions to the audio devices. I'm not
sure whether this is a multi-seat configuration problem or if the culprit is
because synergy doesn't count as a physical pointer in some cases.

This needs some research, I think. But maybe it's even a better idea to run
pulseaudio as a systemd service, because it's a remote sink for dnyarri as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>