Move over from TKabber to Gajim.
This not only patches the desired configuration into Gajim itself but also
enforces a specific GTK+ theme to be used for Gajim only. The main reason for
that is to mimic the look of my previos TKabber theme and configuration.

For the gory details, have a look at the patch. Basically we just use the config
from the nix store as a default configuration file and only add changed lines to
the configuration in the home directory. That way we can still play around with
options while not killing our base configuration. If we want to make something
permanent, we can just move it into the nix config.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>