gajim: Adjust styles of roster and emoticons.
For the emoticons, let's just turn them off. They're annoying as hell!

The rest are just cosmetic changes to get closer to the UI I was familiar with
TKabber.

And well, I had nickname colors in chat windows mixed up so it was a bit hard
for me to follow along (I always had cyan for my own nick and coral for other
nick) because the own vs. other nick colors were swapped.

One particularly annoying thing were those creepy animations of GTK+, so I'm
turning them off. There is still the "smooth scrolling" in chat windows left,
but that's a topic for a commit in the (hopefully not so) distant future :-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>