gajim: Update changed configuration items.
These are the changes I want to make on a permanent basis. The first is to
_always_ "pop up" (it won't _really_ pop up because we're using
single_roster_window) even when dnd or away. The reason is that urgent
notifications aren't very disturbing while "in the zone", but on the other hand
if I really _want_ to know, I want it to be seen even when away.

In addition, this also sets chat state notifications to "composing_only". I'm
sick of messages like "HEY!? Why aren't you paying attention!?" or "Why did you
close the window?".

Signed-off-by: aszlig <aszlig@redmoonstudios.org>