mmrnmhrm: Switch over to "new" (old) hardware.
Unfortunately, the main board has died so I needed to quickly assembly a
replacement machine. I haven't found a compatible main board for the
last set of hardware so the "new" machine is now a Core 2 Duo with only
2 GB of RAM.

I'm looking forward to more frustrations during build I already had with
the i7 I had previously.

Also, all these changes are untested for now, because I'm still
shoveling the btrfs filesystems to two new hard disks, because the new
mainboard only has 2 SATA ports.

At the moment the GRUB bootloader is still on the old disk and as soon
as the data is on the new disks, the GRUB install devices will change as
well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>