i3: Add NixOS options to configure workspaces.
This allows for a more dynamic workspace assignments, especially when
varying between the number of heads. We now not only can use the NixOS
module system to set workspaces but also assign applications to them.

And the default workspace layout is to evenly spread out the heads among
the available heads.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>