Change namespaces into new project name "vuizvui".
I found "nixconf" a little big too generic and thus not unlikely enough
to someday collide with something in <nixpkgs>. That's why I chose the
name "vuizvui" as a Bavarion word (actually _two_ words: "vui zvui") for
"far too much", in terms of the opposite of "nix" - which means
"nothing".

A possible downside for choosing this name is that it might be
jawbreaker to some English native speakers out there, but I don't really
care if the pronunciation is correct nor do I expect to get a lot of
public attention on this repository.

And yes, for English native speakers, a pronunciation like "fui-tsui" is
probably okay as well :-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>