i3: Check whether configuration file is valid.
Finally, after going through quite some journey, we can now check the
config file without needing an X connection. The journey took a bit
longer, because I was stubborn enough to get the test suite running on
Nix (because I wanted to have an upstream test case for this). All that
to create a small patch which just checks that damn config file and
emits an exit code that actually reflects the output :-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>