vim: Move overrides and config into pkgs/vim.
The main reason for doing this is to avoid collisions with other vim
variants in <nixpkgs> which are expecting an .override attribute.

Moving it into pkgs/ entirely also has the advantage of being properly
namespaced rather than "all over the place" and we also don't clash
anymore with existing Vim packages.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>