common-workstation: Remove eye cancer from theme.
Or better: Remove brightness.

I haven't done quite a lot to do make it really nice and shiny. Right
now it only sets a darker background color and color-inverts the panel
image which has the negative side-effect that the NixOS logo is yellow
now, but well... it's just for the login ;-)

Also, this now correctly centers the login on the leftmost screen
instead of centering it across all screens.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>