modules: Add "fancy" vlock-based lock screen.
Unfortunately it's only "fancy" in quotes because the way it's activated
is using a dummy socket file with the user "aszlig" hardcoded at the
moment. In terms of security it isn't a problem, because vlock is using
PAM for authentication and it falls back to authenticating against the
root user.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>