Add a (broken) collection of HumbleBundle games.
This is more or less a bunch of expressions for a few HumbleBundle
games. Right now none of them are working, because fetchHumbleBundle is
broken. But apart from that only ftl and cavestoryplus really work.

The main reason for this commit is to collect those expressions
together, sort them out and refactor them to have a working set of
(maybe not only HumbleBundle) games that are difficult to package in
nixpkgs, because the upstream binaries are not public.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>