pkgs/nixops: Update to latest upstream master.
We also now use fetchFromGitHub and no longer use "git rev-parse" in
order to truncate the commit hash. Instead, we're just trimming down the
revision we already have in the Nix expression to 7 characters (like per
default if done by "git rev-parse --short").

This should also mitigate the evaluation problems with latest nixpkgs,
where leaveDotGit changed so it results in a different hash.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>