tishtushi: Use TLP service instead of hdparm.
Thanks to @abbradar for packaging the TLP nixpkg and NixOS service
module. The upside is that the APM level of the disk is 254 by default,
so we don't need to explicitly set that during bootup and it is done
automatically when we're on AC (when the battery is the power source
it's 128).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>