services_laptop is now split in two
Due to the special TLP service which is only available to Lenovo
ThinkPads I have split services_laptop into services_eris and
services_titan in order to allow a more finegrained control of hardware
specific services. Also, this enables me to use a different font for the
terminal, which really was too small on the ThinkPad.