Add a static configuration for redmoon network.
Today seems to be a "bad luck day" because my Sheeva Plug just died. So
I'm back to some weird stupid old Netgear thingy that seems to be so old
that it doesn't even know what IPv6 "might be". And it's already the
"latest and gro^Heatest" firmware that's available.

One of the ugly things is that our internal DNS configuration can't be
done using that stupid thing, because it uses DHCP to assign external
DNS servers rather than having an internal DNS forwarder where you can
assign your own[TM] zones.

So let's hope that we can revert this in the foreseeable future.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>