tishtushi: Really use the latest mainline kernel.
I can't emphasize this enough: *REALLY*

Whenever the newest stable kernel is released, the version in
linux_testing will *still* the latest -rc even though it's older than
the current stable.

Strictly speaking, it's true that even though the stable version is
newer, the label "testing" is correct. However, and in my case I want to
really have the latest mainline version, no matter whether it's stable
or testing.

So, we're just comparing the major and minor version numbers and if they
match, we're using the latest stable, otherwise we're using testing.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>