bfqsched: Remove cruft for supporting kernel 3.15.
This was in since the initial commit (6c224e4) of the bfqsched Nix
expression and I didn't notice it for all those upgrades.

Basically, this is just a no-op, because nowadays the sed expression
doesn't rewrite anything anymore, but it might be harmful if it should
match something in a future version of the bfqsched patch set.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>