Move common{-workstation}.nix into profiles.
Both files are specific to my machines only and need to be generalized
in order to be useful for anyone else.

Moving these files has a few other censequences, such that we now need
to automatically import the module-list.nix in callMachine.

Speaking of module-list.nix, the file is now alphabetically sorted.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>