machines: Integrate heinrich+labtop from labernix.
So far I'm not quite sure whether mailserver really belongs to labnet,
so I'm leaving it at the labernix subtree. Maybe it even makes sense to
just make it a profile until we have it on a real machine.

We no longer need common.nix for these machines, because it is already
done via callMachine.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>