Merge labernix packages into vuizvui pkgs.
It's currently only BeeHive, but our goal is to dissolve labernix step
by step until nothing is left.

Also, we're now no longer namespace the pkgs with vuizvui directly in
the package list. Before it wasn't even namespaced correctly (except for
inside pkgs/ directly) and we did override the packages using the dirty
approach in overrides/.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>