Move last machine from labernix to vuizvui.
I've moved the restrictions config of Postfix into the default module
for now and actually fixed it so that it's actually working (the config
value wasn't set before). Also, the option type was incorrectly set to
types.list, which aliases to types.listOf and expects another function
(kind) as its argument.

This marks the end of LaberNix and the beginning of a new Vuizvui!

Signed-off-by: aszlig <aszlig@redmoonstudios.org>