services_titan: Make closed lid invoke suspend
This laptop (Toshiba z930) does not go into suspend when the lid is
closed, even though the event is seemingly recognized. To work around
this for the moment, this commit adds an ACPI event command which
listens to the lid's state change and executes pm-suspend accordingly.
Shamelessly ripped from pSub.