tomahawk: Reference quazip library file directly.
It's not a big issue, because due to Nix's stdenv it's going to be able
to link anyway, but there still is a nasty warning, we're getting rid by
this:

WARNING: Target "tomahawklib" requests linking to directory
"/nix/store/ybw942c8jw0dyjah0r0pcsgn8q0sv02h-quazip-0.7.1/lib".  Targets
may link only to libraries.  CMake is dropping the item.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>