labtop: Merge hardware config into one labtop.nix.
This reverts the split up of the config from 55ec3a1 and puts all the
stuff from the hardware configuration file into the main config. The
reason for this is that vuizvui is structured so that the machine
definition *IS* the actual hardware configuration which deviates from a
common profile.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>