release.nix: Include nixpkgs tarball in channel.
Symlinking is actually a pretty bad idea, because we can only have one
channel expression tarball. Evaluation however doesn't realize the
referenced store paths, so we need to include them directly in the
channel tarball.

This also simplifies rewriting of the NIX_PATH references, because we
need to recurse the directory structure only once.

In order to be correctly recognized by nixos-rebuild and display useful
version info, we generate .git-version and .version-suffix now as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>