Handle all <nixpkgs> paths with nixpkgs-path.nix.
This file is just defaulting to <nixpkgs>, but we're going to substitue
it by the channel generator. We also need to make sure that we don't
have any other references to <nixpkgs>, but the latter can best be done
on Hydra's side if we don't make <nixpkgs> available to vuizvui builds.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>