release.nix: Refactor to use patched nixpkgs.
We now use nixpkgsSrc to make sure that Hydra builds will fail as soon
as we hit a reference of <nixpkgs> in either vuizvui of nixpkgs.

This also means, that we don't rewrite all of vuizvui and nixpkgs
anymore and we only really need to rewrite nixos-rebuild.sh.

Although we're using patched nixpkgs for Hydra builds, the release.nix
still has a fallback to <nixpkgs> to make development easier (patching
nixpkgs on every tiny change will take a long time otherwise).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>