release.nix: Fix get-version-suffix in tarball.
This gets rid of an ugly error message during nixos-rebuild, because the
stock get-version-suffix tries to find <nixpkgs>, which will either lead
to the wrong channel (if it even exists at all) or an ugly error.

In both cases however, it doesn't have an effect, because the channel
expression store path is read-only and ships with a .version-suffix
already.

It really only gets rid of ugly error messages.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>