modules/programs: Add module for taskwarrior.
This comes with a patch to chainload taskwarrior configs and defaults to
the config supplied by the module. So if we want to play around with
different configuration values, it's easy to do by just adding a
~/.taskrc.

Note that the patch uses nestlevel 2 for ~/.taskrc, because if we would
use the default (1), the default configuration would be applied prior to
parsing ~/.taskrc, which of course would balk our plan.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>