machines: Allow to easily build ISO images.
I know it's a somewhat hacky approach to strip off "_module",
"boot.loader" and "fileSystems" from the machine config, but that should
be the options which are to be set by iso-image.nix but that way we can
re-use the upstream module.

Also, if one of our modules sets an option without a proper "vuizvui."
namespace, we get an error as well. But it's our policy anyway to always
namespace with "vuizvui." so it's even good to get an error here.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>