machines: Properly override ISO configuration.
We now do it the exact opposite way than introduced in 27dce7b. Instead
of evaluating the machine config and stripping off the options we don't
want to conflict with the iso-image.nix module, we now wrap the
iso-image.nix module itself and just mkForce the values we don't want to
collide.

The reason for this is that the previous implementation just didn't work
because dependent module options from the machine config (for example
config.system.build.*) were already evaluated and thus we end up with
overriding configuration options but get an initrd with the machine
options (which we actually want to override) instead of the
fileSystem/boot options that come with the iso-image.nix module.

Although I'm not quite happy with this approach, it's still better than
the old one and if iso-image.nix gets conflicting options we at least
get a better error message rather than the definitions simply being
stripped off.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>