kzerza: Power off system when GrandPA stops.
This was the behaviour back then when NixOS still used upstart and after
some searching I found that ExecStopPost is what we want here, despite
its name (I'd imply that it is used in parts to stop the service rather
than executed after the service has stopped).

So in turn, it means that using Ctrl-Q within GrandPA shuts down the
whole system.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>