modules/system: Add an option to build an ISO.
Building ISO images for all machines really doesn't make sense if we
only have _one_ machine at the moment, where we really use it (kzerza).

So we now have an option which can be defined for a particular machine
configuration, which is then going te be excempted from the ISO building
process.

The latter isn't yet implemented however, only the option is here right
now.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>