profiles/common: Use nix.nixPath to set NIX_PATH.
We were using environment.sessionVariables before, but
NixOS/nixpkgs@3f1354a added an option exclusively for this. It
internally still sets environment.sessionVariables, so it doesn't change
anything for us because we're overriding the full list anyway.

But should we switch from overriding to appending/prepending someday, we
could get unexpected behaviour, so that's why I'm switching to
nix.nixPath here.

Thanks to @devhell for bringing this to my attention.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>