profiles/common: Force rebuilding GHC+packages.
I was glad that the issues because of the longstanding GHC determinism
bug (https://ghc.haskell.org/trac/ghc/ticket/4012) happened less
frequently with GHC 7.10.

Unfortunately, they still happen, as reported by @Profpatsch at
headcounter/deployment#19 so we're going to override the GHC derivation
with an additional attribute that is exported into the environment of
the builder.

This attribute causes GHC and all dependant packages to be rebuilt on
the Headcounter Hydra. The original issue was because we're using the
official NixOS binary cache as well as building everything from source.

It's helpful for not needing to rebuild the entire world, but for
Haskell packages it's harmful until the mentioned GHC bug has been
fixed.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>