machine_common: Fix inclusion of all lohit fonts.
Since NixOS/nixpkgs@346dcab, lohit-fonts isn't a single derivation
anymore and has been split up into an attrset consisting of the
individual fonts.

So if we want to have all lohit fonts, we need to add all the values of
the attribute set to the list of fonts (and filter out .override and
other functions).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>