tishtushi: Revert patching Dragonfly quirk.
This reverts commit 7f9f88e90b8ab41a97a86fa4ff8a501e0e0eea27.

The reason why I'm reverting this is because while it indeed fixes the
volume mapping, it doesn't fix my main problem that it takes a few
seconds to go from 100% volume to the desired volume level that's
currently set.

So instead of applying this patch and maintaining it until it may
eventually hit mainline, it's better to debug the original issue rather
than applying a patch that _may_ fix an unrelated issue.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>