machines: Patch "nomodeset" out of iso-image.nix.
This is a very hacky approach, because patching that file also requires
us to patch the imports it's referencing.

The first reason why I needed to patch is that there is no "modeset" you
can add to kernelParams and re-enable modesetting that way.

And the second reason is because we don't have something like mkRemove
or mkFilter in the module system, so we could filter out items from a
list.

Another option here would be to mkForce-override the kernelParams, which
would imply that we'd need to duplicate a lot of these options (for
example init=...).

So in the long run we surely need to have a better way to override
lists, but until that I'm leaving it with the patched approach.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>