profiles/common: Pass tests attribute to modules.
This is going to be used in modules or configurations to include tests
specific to the module, for example let's say you have a foo.nix module
that defines a foo service and you have a test named "foo" in the
tests.vuizvui namespace.

You can use it like this:

{ config, tests, lib, ... }:

{
  ... other options of the foo service module ...
  config = lib.mkIf config.foo.enable {
    vuizvui.requiresTests = [ tests.vuizvui.foo ];
  };
}

If a machine config enables the module foo (foo.enable = true) then the
"foo" test is also a constituent of the machine channel for the machine.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>