profiles/common: Move gathering tests to lib/.
So my gut feeling in 90c080e was't wrong after all, so we need to create
jobs for all the tests the machines include.

In order to make this easier, we're going to re-use the collection of
all test attributes from release.nix so putting it in lib/ makes sense.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>