release.nix: Include all tests that are enabled.
This now includes both upstream NixOS and Vuizvui tests.

Unfortunately, evaluation now takes a ton of time and also a ton of
memory, so let's hope our Hydra can sustain it. The reason is that we
need to look through all of the available tests and recurse through
_all_ machine configurations to see if they're defined there.

If it is too heavyweight we might need to find a better solution for
that.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>