profiles/tests: Avoid scanning full linuxPackages.
Using an equality test on an attribute set is actually a very bad idea,
because it forces strict evaluation.

In this case a simple test on the kernel version between the defined one
and the available one should be sufficient.

And if we have a false-positive in test inclusions it doesn't hurt
anyway.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>