aszlig/programs: Add a new "taalo-build" program.
This is primarily for whenever I'm on the road with varying degrees of
internet connectivity.

What it essentially does is ssh to mmrnmhrm, then ssh to taalo and then
run nix-store --serve --write on it. Taalo is the Hydra master of
https://headcounter.org/hydra/ and it has remote builds enabled.

The script essentially only builds on the remote host but doesn't fetch
the builds. The latter can be done if the Hydra is added as a build
cache to the local system, which in case of vuizvui is by default.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>