lib/call-machine: Clean up expression
This gets rid of the use attribute, which is now called "config". We had
the "config" attribute before but it was kinda pointless, because it was
just the import of the path and nothing else.

So the config attribute now is the machine configuration with all of the
vuizvui modules imported as well.

The "build" attribute is now called "eval", which is more appropriate,
because it's the evaluation of the configuration and not the finished
system build.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>