profiles/tests: Add networking.*.loopback test
I've added that test in NixOS/nixpkgs@4c61faa and it fails at the
moment, still leaving the channels in a broken state in regards to
networkd.

The reason I'm doing this anyway is to ensure I won't forget to add that
test here as well. We can't easily go back to the "not broken" state
anymore, except if we roll back to a version with systemd < 228.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>