machines: Add new managed machine "tyree"
Currently this is just a somewhat basic configuration, because the
hardware (an ASUS T100HA) is going to get us in trouble.

For example right now not even the display is working correctly, neither
is WiFi, but we're going to fix that real soon[TM] :-)

The configuration is pretty much based on the "haenk" config, which this
machine will replace once everything is working.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>