profiles/tests: Temporarily disable quake3 test
The Quake 3 demo data have been marked as "unfreeRedistributable" in
NixOS/nixpkgs@6b447a3, which causes this test to not evaluate unless we
set config.allowUnfree to true.

I guess the best solution to fix this upstream is to use something like
OpenArena for the Quake 3 tests, because I didn't manage to find
anything regarding the demo files and/or whether those are
redistributable or not.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>