humblebundle: Add GuacaMelee
This comes with an LD_PRELOAD wrapper that is just used for the initial
chdir() where the game tries to set its working directory to the data
files.

After that all loading of the preload wrapper will fail, but that's
okay because we just need to set the data directory and be done with it.

We also introduce a callPackage_i686 which is equivalent to the one in
<nixpkgs> but also adds our own packages to the set.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>