pkgs/axbo: Re-package with open source code
So far I've used the proprietary download and patched it up to work on
NixOS. This now goes through the hurdles of trying to package it by
building it from source, although there are still some precompiled
dependencies used:

  commons-beanutils-1.8.3
  commons-digester3-3.2
  jcommon-1.0.20
  jfreechart-1.0.16
  swingx-all-1.6.4

In order to push this into upstream <nixpkgs> we first need to package
these.

Another issue which I didn't dig into is that the interface now looks
quite ugly. I'm not sure whether it is because of some packaging choices
or because we're now at version 3.x and it's supposed to look like this.

Nevertheless, the application works and it makes more sense to improve
the source package than the patched up binary package.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>