Add support for fetching non-DRM Steam games
Thanks to the folks at @SteamRE there is a reverse-engineered Steam
library and a downloader for Steam depots.

The downloader actually uses mcbyte-it/DepotDownloader@5fa6621 which is
a fork that has added proper exit codes.

We also patch the downloader to show the latest manifest ID, so we can
check whether we're up to date.

The reason we're pinning the manifest ID is that we make sure that the
SHA256 will always match, no matter whether there is a new upstream
version or not.

Obviously the whole steam/ namespace is only for Steam games that can
run without Steam, which in turn is the entire purpose of it.

For a list of Steam games without DRM, have a look at:

http://steam.wikia.com/wiki/List_of_DRM-free_games

Also if you want to look up the depotId or appId, this is a good
resource:

https://steamdb.info/

Signed-off-by: aszlig <aszlig@redmoonstudios.org>