steam: Add Starbound
Comes in two flavors: stable (beta) and unstable (beta), though at the
moment, both are the same version.

I've renamed prefixed the command names with starbound-... and also
converted underscores to dashes to make the commands a bit more
consistent.

Also, I made room for Mac OS X and Windows so in *theory* it should also
work on these platforms. However I only tested it on NixOS with a 64bit
Linux.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>