mmrnmhrm/dnyarri: Update kernel src/cfg to 4.5-rc4
Geesh, the configuration is getting more and more rotten and it's time
to make this in a more generic configuration *very* soon.

The configuration does have a lot of cruft in it because it's a bunch of
"make oldconfig" iterations and no cleanup in-between.

In addition, even if I'd do the cleanup I'd probably want common options
to be factored out.

But for now let's keep the config as-is until 4.6 comes out and we
either play the "make oldconfig" game again or we finally rewrite it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>