hardware/t100ha: Add custom kernel + patches
The custom kernel has built in modules for the MMC storage and a small
patch which is essentially diff of the merge of the "drm-intel-fixes"
branch from git://anongit.freedesktop.org/drm-intel against 4.5-rc5.

This is the HEAD of what I have merged:

https://cgit.freedesktop.org/drm-intel/commit/?h=drm-intel-fixes&id=9b18572e83bfd5378b3fcff3acf123f7bddf558a

Eventually these fixes will hit mainline so we can drop them very soon.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>