manual: Fix evaluation error
Since NixOS/nixpkgs@cad8957 there is an additional argument to
optionAttrSetToDocList, so we need to pass internal _module arguments to
that argument (typically _module.check and _module.typeInference) which
are used by the type checker to find errors.

On Vuizvui however, we don't build the full manual but only
Vuizvui-specific options, so we might want to aim for a different
approach in the long term rather than duplicating a lot of the manual
generation code.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>