hardware/t100ha: Switch to linux-next with patches
Both of these patches are from the following Google Drive:

https://drive.google.com/folderview?id=0B4DiU2o72FbuejQ1S2VZZW5xV2c

The meta-keys-asus.patch (which is called just meta-keys.patch in this
repository) is used verbatim.

However, the baytrail-backlight.4.4.patch (here just backlight.patch) is
a rebased version of the patch from the Google Drive with a few
modifications by me, which boils down to:

 * Remove the module parameter force_backlight_pmic, because this module
   is hardware-specific so we don't actually need to do that (and *if*
   we want to do that we'd compile it in directly).
 * Add an unused pipe function argument to vlv_pmic_setup_backlight().

As the backlight patch introduces the functions intel_soc_pmic_readb()
and intel_soc_pmic_writeb() which are not available at module link time,
I have also added DRM_I915 to be compiled into the kernel.

In addition I needed to disable VIDEO_EM28XX and RAPIDIO, because they
do not compile with linux-next-20160226 and I didn't bother to provide
fixes because it's for hardware that is not existing on the T100HA.

Note that I'm using linux-next-20160226 here instead of 20160229 because
the latter has some networking I/O issues right now.

This makes the backlight, battery status and charging usable on the
T100HA and the fixes from the drm-intel-fixes branch are no longer
needed because they're already in linux-next-20160229.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>