rtl8192cu: Move enableRTL8192cFirmware from tyree
This really is specific to the RTL8192CX chipset and I originally was
using this for tyree along with the mainline kernel modules.

Unfortunately, I had connection drop outs so I switched to the original
Realtek driver, which is now vuizvui.hardware.rtl8192cu, so the firmware
is used and should be enabled there as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>