games/starbound: Refactor to use JQ for bootconfig
While JQ is a beast of its own implementing its own turing complete
language, it *is* certainly useful and better than sed to replace path
values.

Now we have a new bootOverrides attribute set, which is converted into
JSON and then merged with the original sbboot.config by JQ at the time
we're setting up the XDG_DATA_HOME directory.

The reason this is necessary is that we can't use environment variables
directly in the boot configuration of Starbound, so we need to write the
boot configuration at initialization time with the absolute paths to
XDG_DATA_HOME.

Another change here is that we now reference the upstream store path
directly instead of copying over almost everything to our build result.

This has the advantage that we have fewer duplication and thus fewer
disk I/O while building.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>