modules: Add new Starbound service and test
Very preliminary and doesn't have all the option descriptions right, nor
does it have convenience features such as setting allowAdminCommands
based on whether any users are defined with admin privileges.

Of course the latter needs to undergo the decision on how to handle RCON
connections, because the latter *might* need that option.

But apart from that single option, there are a lot more options we need
to flesh out.

Also, the test currently is very limited and only spins up a client,
connects to the server and does a movement (just walk to the right).

Needless to say, it's even quite fragile and relies on OCR to properly
detect the custom pixel fonts from Starbound. Which unfortunately fails
most of the time.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>