Add a new module and test for gpg-agent
Since NixOS/nixpkgs@5391882 there no longer is the option to start the
agent during X session startup, which prompted me to write this module.

I was unhappy how GnuPG is handled in NixOS since a long time and wanted
to OCD all the configuration files directly into the module.

Unfortunately, this is something I eventually gave up because GnuPG's
design makes it very hard to preseed configuration. My first attempt was
to provide default configuration files in /etc/gnupg, but that wasn't
properly picked up by GnuPG.

Another way would have been to change the default configuration files,
but that would have the downside that we could only override those
configurations using command line options for each individual GnuPG
component.

The approach I tried to go for was to patch GnuPG so that all the
defaults are directly set in the source code using a giant sed
expression. It turned out that this approach doesn't work very well,
because every component has implemented its own ways how to handle
commandline arguments versus (default) configuration files.

In the end I gave up trying to OCD anything related to GnuPG
configuration and concentrated just on the agent.

And that's another beast, which unfortunately doesn't work very well
with systemd.

While searching the net for existing patches I stumbled upon one done by
@shlevy:

https://lists.gnupg.org/pipermail/gnupg-devel/2014-November/029092.html

Unfortunately, the upstream author seems to be quite anti-systemd and
didn't want to accept that into the upstream project.

Because of this I went for using LD_PRELOAD to pick up the file
descriptors provided by the systemd sockets, because in the end I don't
want to constantly catch up with upstream and rebase the patch on every
new release.

Apart from just wrapping the agent to be socket activated, we also wrap
the pinentry program, so that we can inject a _CLIENT_PID environment
variable from the LD_PRELOAD wrapper that is picked up by the pinentry
wrapper to determine the TTY and/or display of the client communicating
with the agent.

The wrapper uses the proc filesystem to get all the relevant information
and passes it to the real pinentry.

The advantage of this is that we don't need to do things such as
"gpg-connect-agent updatestartuptty /bye" or any other workarounds and
even if we connect via SSH the agent should be able to correctly pick up
the TTY and/or display.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>