modules/gpg-agent: Hide internals of preloader
First of all, let's inline the first_fork variable, because we don't
want this variable to be exposed as a symbol, even though it doesn't
hurt (except maybe for a very very very tiny improvement in RTLD lookup
performance).

And apart from the first_fork variable, there were a few other symbols
we don't want to propagate to the RTLD chain as well:

0000000000001465 T accept
000000000000130b T bind
000000000000153e T execv
0000000000001610 T _fini
00000000000013b8 T fork
0000000000000fe0 T get_sd_fd_for
0000000000001420 T get_socket_pid
0000000000000d80 T _init
00000000000012fa T listen
00000000000012b8 T record_sockfd

So in the end we're down to:

00000000000011fb T accept
00000000000010a1 T bind
00000000000012c8 T execv
0000000000001390 T _fini
000000000000114e T fork
0000000000000b68 T _init
0000000000001090 T listen

... which is a lot more clean and even though our staff doesn't collide
with existing libraries in the chain it's better to be safe than sorry.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>